+++ tests/estimators/test_functional.py
@@ -69,8 +69,8 @@ def test_entropy_functional_addressing(entropy_approach):
     approach_str, needed_kwargs = entropy_approach
     data = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10]
 
-    # ANSB returns NaN for data without coincidences
-    if approach_str == "ansb":
+    # ANSB and NSB return NaN for data without coincidences
+    if approach_str in ["ansb", "nsb"]:
         entropy = im.entropy(np.array(data), approach=approach_str, **needed_kwargs)
         assert np.isnan(entropy)
         # test list input
@@ -91,8 +91,8 @@ def test_entropy_class_addressing(entropy_approach):
     est = im.estimator(data, measure="entropy", approach=approach_str, **needed_kwargs)
     assert isinstance(est, EntropyEstimator)
 
-    # ANSB returns NaN for data without coincidences
-    if approach_str == "ansb":
+    # ANSB and NSB return NaN for data without coincidences
+    if approach_str in ["ansb", "nsb"]:
         assert np.isnan(est.result())
         assert np.isnan(est.global_val())
     else:
@@ -103,7 +103,7 @@ def test_entropy_class_addressing(entropy_approach):
     if approach_str in ["renyi", "tsallis", "chao_shen", "cs", "bayes"]:
         with pytest.raises(UnsupportedOperation):
             est.local_vals()
-    elif approach_str in ["chao_wang_jost", "cwj", "ansb"]:
+    elif approach_str in ["chao_wang_jost", "cwj", "ansb", "nsb"]:
         with pytest.raises(TheoreticalInconsistencyError):
             est.local_vals()
     else:
@@ -125,6 +125,7 @@ def test_cross_entropy_functional_addressing(entropy_approach, default_rng):
         "chao_wang_jost",
         "cwj",
         "ansb",
+        "nsb",
     ]:
         with pytest.raises(TheoreticalInconsistencyError):
             im.entropy(data, data, approach=approach_str, **needed_kwargs)
@@ -163,6 +164,7 @@ def test_cross_entropy_class_addressing(entropy_approach, default_rng):
         "chao_wang_jost",
         "cwj",
         "ansb",
+        "nsb",
     ]:
         with pytest.raises(TheoreticalInconsistencyError):
             est = im.estimator(
@@ -272,6 +274,7 @@ def test_cross_entropy_functional_random_symmetry(entropy_approach, default_rng)
         "chao_wang_jost",
         "cwj",
         "ansb",
+        "nsb",
     ]:
         with pytest.raises(TheoreticalInconsistencyError):
             im.cross_entropy(p, q, approach=approach_str, **needed_kwargs)